Restructure modules into package
