+++ runepy/__init__.py
@@ -1,0 +1,6 @@
+"""RunePy package."""
+
+from . import pathfinding
+
+__all__ = ["pathfinding"]
+
