+++ runepy/utils.py
